refactored share packages
