+++ app/src/main/AndroidManifest.xml
@@ -24,7 +24,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".sharetwitter.LoginActivity"
+            android:name=".share.twitter.LoginActivity"
             android:screenOrientation="portrait">
             <intent-filter>
                 <action android:name="android.intent.action.VIEW" />
@@ -305,7 +305,7 @@
             android:exported="true" />
 
         <activity
-            android:name=".SharingActivity"
+            android:name=".share.SharingActivity"
             android:screenOrientation="portrait"
             android:theme="@style/AppTheme.NoActionBar" />
 
@@ -332,19 +332,19 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".sharedrupal.DrupalLogin"
+            android:name=".share.drupal.DrupalLogin"
             android:theme="@style/Theme.AppCompat.NoActionBar" />
         <activity
-            android:name=".sharewordpress.WordpressLoginActivity"
+            android:name=".share.wordpress.WordpressLoginActivity"
             android:theme="@style/Theme.AppCompat.NoActionBar"/>
 
-        <activity android:name=".nextcloud.NextCloudAuth"
+        <activity android:name=".share.nextcloud.NextCloudAuth"
             android:theme="@style/AppTheme.NoActionBar"/>
 
 
 
         <activity
-            android:name=".sharetoflickr.FlickrActivity"
+            android:name=".share.flickr.FlickrActivity"
             android:label="@string/app_name"
             android:theme="@style/Theme.AppCompat.NoActionBar"
             android:launchMode="singleTask" >
@@ -355,11 +355,11 @@
                 <data android:scheme="flickrj-android-sample-oauth" />
             </intent-filter>
         </activity>
-        <activity android:name=".sharedimgur.ImgurAuthActivity"
+        <activity android:name=".share.imgur.ImgurAuthActivity"
             android:theme="@style/AppTheme.NoActionBar"/>
 
         <activity
-            android:name=".share.shareowncloud.OwnCloudActivity"
+            android:name=".share.owncloud.OwnCloudActivity"
             android:label="@string/title_activity_own_cloud"
             android:theme="@style/AppTheme.NoActionBar"></activity>
         <meta-data
