+++ app/src/main/java/org/fossasia/phimpme/accounts/AccountActivity.java
@@ -51,12 +51,12 @@ import org.fossasia.phimpme.base.RecyclerItemClickListner;
 import org.fossasia.phimpme.base.ThemedActivity;
 import org.fossasia.phimpme.data.local.AccountDatabase;
 import org.fossasia.phimpme.data.local.DatabaseHelper;
-import org.fossasia.phimpme.nextcloud.NextCloudAuth;
-import org.fossasia.phimpme.share.shareowncloud.OwnCloudActivity;
-import org.fossasia.phimpme.sharedimgur.ImgurAuthActivity;
-import org.fossasia.phimpme.sharedrupal.DrupalLogin;
-import org.fossasia.phimpme.sharetoflickr.FlickrActivity;
-import org.fossasia.phimpme.sharewordpress.WordpressLoginActivity;
+import org.fossasia.phimpme.share.nextcloud.NextCloudAuth;
+import org.fossasia.phimpme.share.owncloud.OwnCloudActivity;
+import org.fossasia.phimpme.share.imgur.ImgurAuthActivity;
+import org.fossasia.phimpme.share.drupal.DrupalLogin;
+import org.fossasia.phimpme.share.flickr.FlickrActivity;
+import org.fossasia.phimpme.share.wordpress.WordpressLoginActivity;
 import org.fossasia.phimpme.utilities.ActivitySwitchHelper;
 import org.fossasia.phimpme.utilities.BasicCallBack;
 import org.fossasia.phimpme.utilities.SnackBarHandler;
