+++ app/src/main/java/org/fossasia/phimpme/leafpic/activities/SingleMediaActivity.java
@@ -46,7 +46,7 @@ import com.mikepenz.community_material_typeface_library.CommunityMaterial;
 import com.yalantis.ucrop.UCrop;
 
 import org.fossasia.phimpme.R;
-import org.fossasia.phimpme.SharingActivity;
+import org.fossasia.phimpme.share.SharingActivity;
 import org.fossasia.phimpme.base.SharedMediaActivity;
 import org.fossasia.phimpme.base.ThemedActivity;
 import org.fossasia.phimpme.data.local.DatabaseHelper;
