+++ app/src/main/java/org/fossasia/phimpme/opencamera/Camera/PhotoActivity.java
@@ -25,12 +25,12 @@ import org.fossasia.phimpme.R;
 import org.fossasia.phimpme.base.ThemedActivity;
 import org.fossasia.phimpme.editor.FileUtils;
 import org.fossasia.phimpme.editor.EditImageActivity;
-import org.fossasia.phimpme.SharingActivity;
+import org.fossasia.phimpme.share.SharingActivity;
 import org.fossasia.phimpme.utilities.ActivitySwitchHelper;
 
 import uk.co.senab.photoview.PhotoView;
 
-import static org.fossasia.phimpme.SharingActivity.EXTRA_OUTPUT;
+import static org.fossasia.phimpme.share.SharingActivity.EXTRA_OUTPUT;
 
 
 public class PhotoActivity extends ThemedActivity {
