+++ app/src/main/java/org/fossasia/phimpme/share/drupal/DrupalLogin.java
@@ -1,7 +1,6 @@
-package org.fossasia.phimpme.sharedrupal;
+package org.fossasia.phimpme.share.drupal;
 
 import android.support.design.widget.Snackbar;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.support.v7.widget.Toolbar;
 import android.view.View;
