+++ app/src/main/java/org/fossasia/phimpme/share/flickr/tasks/GetOAuthTokenTask.java
@@ -1,8 +1,8 @@
-package org.fossasia.phimpme.sharetoflickr.tasks;
+package org.fossasia.phimpme.share.flickr.tasks;
 
 import android.os.AsyncTask;
-import org.fossasia.phimpme.sharetoflickr.FlickrHelper;
-import org.fossasia.phimpme.sharetoflickr.FlickrActivity;
+import org.fossasia.phimpme.share.flickr.FlickrHelper;
+import org.fossasia.phimpme.share.flickr.FlickrActivity;
 import com.googlecode.flickrjandroid.Flickr;
 import com.googlecode.flickrjandroid.oauth.OAuth;
 import com.googlecode.flickrjandroid.oauth.OAuthInterface;
