+++ app/src/main/java/org/fossasia/phimpme/share/nextcloud/NextCloudAuth.java
@@ -1,10 +1,9 @@
-package org.fossasia.phimpme.share.shareowncloud;
+package org.fossasia.phimpme.share.nextcloud;
 
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
 import android.support.design.widget.TextInputLayout;
-import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
@@ -14,11 +13,12 @@ import com.owncloud.android.lib.common.OwnCloudClientFactory;
 import com.owncloud.android.lib.common.OwnCloudCredentialsFactory;
 
 import org.fossasia.phimpme.R;
+import org.fossasia.phimpme.base.ThemedActivity;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
 
-public class OwnCloudActivity extends AppCompatActivity implements Button.OnClickListener  {
+public class NextCloudAuth extends ThemedActivity implements Button.OnClickListener {
 
     private static final int RESULT_OK = 1;
 
@@ -43,13 +43,14 @@ public class OwnCloudActivity extends AppCompatActivity implements Button.OnClic
     @BindView(R.id.buttonOK)
     Button buttonOK;
 
+
     private OwnCloudClient authClient;
     private String username, password;
 
     @Override
-    protected void onCreate(Bundle savedInstanceState) {
+    public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_own_cloud);
+        setContentView(R.layout.activity_next_cloud_auth);
         ButterKnife.bind(this);
         username = accountUsername.getText().toString();
         password = accountUsername.getText().toString();
@@ -67,6 +68,8 @@ public class OwnCloudActivity extends AppCompatActivity implements Button.OnClic
         } else{
             Intent loginIntent = new Intent();
             Uri serverUri = Uri.parse(hostUrlInput.getText().toString().trim());
+            loginIntent.putExtra(getString(R.string.server_url)
+                    , hostUrlInput.getText().toString().trim());
             loginIntent.putExtra(getString(R.string.auth_username),
                     accountUsername.getText().toString().trim());
             loginIntent.putExtra(getString(R.string.auth_password),
