+++ app/src/main/java/org/fossasia/phimpme/share/owncloud/OwnCloudActivity.java
@@ -1,9 +1,10 @@
-package org.fossasia.phimpme.nextcloud;
+package org.fossasia.phimpme.share.owncloud;
 
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
 import android.support.design.widget.TextInputLayout;
+import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
@@ -13,12 +14,11 @@ import com.owncloud.android.lib.common.OwnCloudClientFactory;
 import com.owncloud.android.lib.common.OwnCloudCredentialsFactory;
 
 import org.fossasia.phimpme.R;
-import org.fossasia.phimpme.base.ThemedActivity;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
 
-public class NextCloudAuth extends ThemedActivity implements Button.OnClickListener {
+public class OwnCloudActivity extends AppCompatActivity implements Button.OnClickListener  {
 
     private static final int RESULT_OK = 1;
 
@@ -43,14 +43,13 @@ public class NextCloudAuth extends ThemedActivity implements Button.OnClickListe
     @BindView(R.id.buttonOK)
     Button buttonOK;
 
-
     private OwnCloudClient authClient;
     private String username, password;
 
     @Override
-    public void onCreate(Bundle savedInstanceState) {
+    protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_next_cloud_auth);
+        setContentView(R.layout.activity_own_cloud);
         ButterKnife.bind(this);
         username = accountUsername.getText().toString();
         password = accountUsername.getText().toString();
@@ -68,8 +67,6 @@ public class NextCloudAuth extends ThemedActivity implements Button.OnClickListe
         } else{
             Intent loginIntent = new Intent();
             Uri serverUri = Uri.parse(hostUrlInput.getText().toString().trim());
-            loginIntent.putExtra(getString(R.string.server_url)
-                    , hostUrlInput.getText().toString().trim());
             loginIntent.putExtra(getString(R.string.auth_username),
                     accountUsername.getText().toString().trim());
             loginIntent.putExtra(getString(R.string.auth_password),
