+++ app/src/main/java/org/fossasia/phimpme/share/twitter/HelperMethods.java
@@ -13,9 +13,8 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.fossasia.phimpme.sharetwitter;
+package org.fossasia.phimpme.share.twitter;
 
-import android.app.Activity;
 import android.content.Context;
 import android.util.Log;
