+++ app/src/main/java/org/fossasia/phimpme/share/twitter/LoginActivity.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.fossasia.phimpme.sharetwitter;
+package org.fossasia.phimpme.share.twitter;
 
 import android.content.Context;
 import android.content.Intent;
