+++ app/src/main/java/org/fossasia/phimpme/share/wordpress/WordpressLoginActivity.java
@@ -1,4 +1,4 @@
-package org.fossasia.phimpme.sharewordpress;
+package org.fossasia.phimpme.share.wordpress;
 
 import android.os.Bundle;
 import android.support.v7.widget.Toolbar;
